+++ src/main/java/net/p3pp3rf1y/sophisticatedbackpacks/compat/mkb/ModernKeyBindingCompat.java
@@ -5,7 +5,11 @@ import committee.nova.mkb.api.IKeyBinding;
 import committee.nova.mkb.keybinding.KeyConflictContext;
 import committee.nova.mkb.keybinding.KeyModifier;
 
+import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
+import net.fabricmc.fabric.api.client.screen.v1.ScreenKeyboardEvents;
+import net.fabricmc.fabric.api.client.screen.v1.ScreenMouseEvents;
 import net.minecraft.client.KeyMapping;
+import net.minecraft.client.gui.screens.Screen;
 import net.p3pp3rf1y.sophisticatedcore.compat.ICompat;
 
 import static com.mojang.blaze3d.platform.InputConstants.KEY_X;
@@ -17,6 +21,8 @@ import static net.p3pp3rf1y.sophisticatedbackpacks.client.KeybindHandler.INVENTO
 import static net.p3pp3rf1y.sophisticatedbackpacks.client.KeybindHandler.SORT_KEYBIND;
 import static net.p3pp3rf1y.sophisticatedbackpacks.client.KeybindHandler.TOOL_SWAP_KEYBIND;
 import static net.p3pp3rf1y.sophisticatedbackpacks.client.KeybindHandler.UPGRADE_SLOT_TOGGLE_KEYBINDS;
+import static net.p3pp3rf1y.sophisticatedbackpacks.client.KeybindHandler.sendBackpackOpenOrCloseMessage;
+import static net.p3pp3rf1y.sophisticatedbackpacks.client.KeybindHandler.tryCallSort;
 
 public class ModernKeyBindingCompat implements ICompat {
 	@Override
@@ -34,5 +40,32 @@ public class ModernKeyBindingCompat implements ICompat {
 		((IKeyBinding) BACKPACK_TOGGLE_UPGRADE_2).setKeyModifierAndCode(KeyModifier.ALT, InputConstants.Type.KEYSYM.getOrCreate(KEY_X));
 
 		UPGRADE_SLOT_TOGGLE_KEYBINDS.forEach((slot, keybind) -> ((IKeyBinding)keybind).setKeyConflictContext(KeyConflictContext.UNIVERSAL));
+
+		ScreenEvents.BEFORE_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
+			ScreenKeyboardEvents.allowKeyPress(screen).register(ModernKeyBindingCompat::handleGuiKeyPress);
+			ScreenMouseEvents.allowMouseClick(screen).register(ModernKeyBindingCompat::handleGuiMouseKeyPress);
+		});
+	}
+
+	public static boolean handleGuiKeyPress(Screen screen, int keycode, int scancode, int modifiers) {
+		InputConstants.Key key = InputConstants.getKey(keycode, scancode);
+		if (((IKeyBinding) SORT_KEYBIND).isActiveAndMatches(key) && tryCallSort(screen)) {
+			return false;
+		} else if (((IKeyBinding) BACKPACK_OPEN_KEYBIND).isActiveAndMatches(key) && sendBackpackOpenOrCloseMessage()) {
+			return false;
+		}
+
+		return true;
+	}
+
+	public static boolean handleGuiMouseKeyPress(Screen screen, double mouseX, double mouseY, int button) {
+		InputConstants.Key key = InputConstants.Type.MOUSE.getOrCreate(button);
+		if (((IKeyBinding) SORT_KEYBIND).isActiveAndMatches(key) && tryCallSort(screen)) {
+			return false;
+		} else if (((IKeyBinding) BACKPACK_OPEN_KEYBIND).isActiveAndMatches(key) && sendBackpackOpenOrCloseMessage()) {
+			return false;
+		}
+
+		return true;
 	}
 }