+++ changelog.md
@@ -7,3 +7,4 @@
 - 
 ### Fixes
 - Fix a compatibility issue with repurposed structures
+- Sorting keybind was not working with keys only with mouse buttons